Initialize WebserverTestUtils properly
